Move the Operation extension to Operations.kt
